pyproject.toml
[tool.poetry]
name = "pywright"
version = "0.1.0"
description = ""
authors = ["pjfig <pjsfigueira@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pytest-playwright = "^0.4.4"
pre-commit = "^3.6.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
